requirements.txt
# Gestión de Liquidación - Requerimientos

# Framework Web
streamlit>=1.31.0

# Base de datos
mysql-connector-python>=8.3.0

# Manejo de Datos
pandas>=2.2.0
openpyxl>=3.1.2

# Utilidades
uuid

# Generación de PDFs
fpdf2>=2.7.8

# Visualización
plotly>=5.18.0
